# perf: rework decoder interface