… src/encoding.zig